Crate nbstr

A lean Cow<'static, str> that cannot be written to.

Size is reduced by replacing String with Box<str>, which removes oneusize`, at the cost of some reallocation.

To avoid boxing when possible, short strs can be stored inside the struct itself, replacing pointer and length. The length of the short str is then stored as a part of the tag/discriminant, which is why Nbstr is a struct and not an enum. The definition of 'short' depends on architecture and features.

There are four variants of nbstr; See README for details.

Examples

extern crate nbstr;
use nbstr::Nbstr;

#[derive(Default)]
struct Container {// <- no lifetime
    list: Vec<Nbstr>
}
impl Container {
    fn append<S:Into<Nbstr>>(&mut self,  s: S) {
        self.list.push(s.into());
    }
}
fn main() {
    let mut c = Container::default();
    c.append("foo");// &'static str
    {   // &str wouldn't work here since the strings goes out of scope before the Vec
        c.append(Nbstr::from_str(&("bar".to_string())));// is short enough to avoid allocating,
        c.append("baz".to_string());
    }
    println!("{:?}", c.list);
}

Structs

Nbstr
A lean Cow<'static, str> that cannot be written to.